+++ CHANGELOG.md
@@ -14,6 +14,7 @@ Note that the `[Unreleased]` section contains all changes that haven't yet made 
 ### Changed
 - Show a warning when using Arch Linux styled architectures (those can be seen from the output of `uname -p`). Such architectures are still allowed for builds, but makedeb prompts the user to switch the architecture listing in the PKGBUILD to the Debian styled one (which can be seen from the output of `dpkg --print-architecture`).
 - Updated `CONTRIBUTING.md` with better guidelines.
+- Require `pkgdesc` to not be empty and contain characters other than spaces when variable is declared in a PKGBUILD.
 
 ### Deprecated
 - An architecture listing of `any` now sets the package architecture to the system's DPKG architecture (as required by the [Debian control file specification](https://www.debian.org/doc/debian-policy/ch-controlfields.html#s-f-architecture)). Previously `any` was automatically converted to `all` (which identifies an architecture-independent package), and using `all` now fulfills that purpose.